fix(static): R7.1–7.3 housekeeping — soft_max clamp, GeoJSON validation, per-index hazard hysteresis
Static-branch review batch (docs/superpowers/plans/2026-07-02-static-branch-review-fixes.md):

- R7.1: clamp StaticObstacleParams::soft_max to ≤0.9 in the model constructor so
  the keep-clear buffer stays strictly below the tracker's 0.95 hard gate — a
  soft_max ≥0.95 would silently turn the soft buffer into a hard no-birth zone.
- R7.2: GeoJsonStaticObstacles now skips (not throws) features with non-numeric
  or out-of-range (lat∉[-90,90], lon∉[-180,180]) coordinates or negative radii,
  and wraps malformed JSON in the documented std::runtime_error instead of
  leaking nlohmann::json::parse_error. Header contract updated.
- R7.3: StaticHazardEvaluator keys hysteresis state by obstacle index, not
  hazard_id, so co-located ENC records (colliding id) keep independent
  enter/exit state; staticHazardId also mixes source_id when present to
  disambiguate them. Events still carry hazard_id.

Pure correctness — no bench-metric change (soft_max default already 0.9; GeoJSON
change only affects malformed input; hazard id/hysteresis do not feed tracking
metrics). 10 new tests; full suite 868 green.

+++ core/collision/StaticHazardEvaluator.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cstdint>
+#include <cstddef>
 #include <map>
 
 #include <Eigen/Core>
@@ -43,7 +43,11 @@ class StaticHazardEvaluator {
   const IStaticObstacleModel* model_{nullptr};
   IStaticHazardSink* sink_{nullptr};
   Config cfg_;
-  std::map<std::uint64_t, bool> inside_;  // hazard_id -> currently inside ring
+  // R7.3: keyed by obstacle index (stable within a run), NOT hazard_id — two
+  // co-located ENC records share a hazard_id, and sharing hysteresis state
+  // there lets one obstacle's Exit mask another's Enter. Events still carry
+  // hazard_id for operator identification.
+  std::map<std::size_t, bool> inside_;  // obstacle index -> inside ring
 };
 
 }  // namespace navtracker
